backend/ML/requirements_ocr.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
google-generativeai>=0.3.0
pillow>=10.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
